+++ CHANGELOG.md
@@ -67,3 +67,4 @@
 - book_rarity added to skill_table. it's used during book creation in db.c file.
 - act_color() implemented, now act_new() calls act_color().
 - fight messages colored and dam count is given for each hit.
+- sector and area info added to room name. (A Room in Mud School [City] [mud school])